Extract bulk operation creatation from databaseStream
